backend/requirements.txt
Django==5.2.12
djangorestframework==3.16.1
psycopg2-binary==2.9.11
django-cors-headers==4.3.1
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
requests==2.32.3
